+++ pytest.ini
@@ -1,8 +1,2 @@
 [pytest]
 DJANGO_SETTINGS_MODULE = core.settings
-python_files = test_*.py
-
-markers =
-    slow: marks tests as slow (deselect with '-m "not slow"')
-    serial
-
