+++ requirements.txt
@@ -127,3 +127,5 @@ xlwt==1.3.0
 xmltodict==0.13.0
 zipp==3.11.0
 ruamel.yaml
+testcontainers
+sqlalchemy
